# api/openapi.yaml
openapi: 3.1.0
info:
  title: AURA — Ashesi University Resource Allocation API
  version: "1.0.0"
  description: >
    Contract-first REST API for AURA, Ashesi University's resource allocation and
    reservation platform (§8). JSON over HTTPS,
    versioned under /api/v1. Errors use application/problem+json (RFC 9457) with a
    stable machine `code`. Cursor pagination via ?limit & ?cursor. This document is
    the source of truth from which the web and mobile typed clients are generated.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.aura.ashesi.edu
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/bookings:
    post:
      tags: [Bookings]
      summary: Submit a booking request (FR7)
      parameters:
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [room_id, purpose, attendee_count, starts_at, ends_at]
              properties:
                room_id: { type: string, format: uuid }
                purpose: { type: string }
                attendee_count: { type: integer, minimum: 1 }
                starts_at: { type: string, format: date-time }
                ends_at: { type: string, format: date-time }
      responses:
        "201":
          description: Created (PENDING)
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking: { $ref: "#/components/schemas/Booking" }
                  competing_pending: { type: integer }
        "422": { $ref: "#/components/responses/Problem", description: "Conflicts with lecture/maintenance, capacity, or past" }
    get:
      tags: [Bookings]
      summary: List bookings (scope mine|pending|all)
      description: >
        Returns a cursor-paginated {data, next_cursor} envelope. For scope=pending
        (the approvals queue), each data item is a BookingApprovability object —
        the booking enriched with nested room/requester plus the §11/FR8 blockers
        explaining whether it can be approved. For scope=mine|all, each data item
        is a plain Booking.
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, pending, all] } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/BookingStatus" } }
        - { name: room_id, in: query, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: >
            OK. data is an array of Booking (scope mine|all) or
            BookingApprovability (scope pending).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      oneOf:
                        - { $ref: "#/components/schemas/Booking" }
                        - { $ref: "#/components/schemas/BookingApprovability" }
                  next_cursor: { type: string, nullable: true }
...
  /api/v1/bookings/{id}:
    get:
      tags: [Bookings]
      summary: Booking detail (owner or BOOKING_OFFICER)
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Problem" } }
  /api/v1/bookings/{id}/approve:
    post:
      tags: [Bookings]
      summary: Approve (BOOKING_OFFICER) — atomic, concurrency-safe (§7.3)
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: Approved }
        "409": { $ref: "#/components/responses/Problem", description: "SLOT_NO_LONGER_AVAILABLE" }
  /api/v1/bookings/{id}/reject:
    post:
      tags: [Bookings]
      summary: Reject with a required note (BOOKING_OFFICER)
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody: { content: { application/json: { schema: { type: object, required: [note], properties: { note: { type: string } } } } } }
      responses: { "200": { description: Rejected } }
  /api/v1/bookings/{id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel (owner or BOOKING_OFFICER)
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses: { "200": { description: Cancelled } }
  /api/v1/bookings/{id}/override:
    post:
      tags: [Bookings]
      summary: Admin override (BR6) — force approve, cancelling conflicts
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses: { "200": { description: Overridden } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    cookieAuth: { type: apiKey, in: cookie, name: __Host-access }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Limit: { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
  responses:
    Problem:
      description: RFC 9457 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Stable machine-readable error code }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
    Role: { type: string, enum: [SYSTEM_ADMIN, TIMETABLE_ADMIN, BOOKING_OFFICER, REQUESTER] }
    RoomType: { type: string, enum: [LECTURE_HALL, LAB, SEMINAR_ROOM, AUDITORIUM, CONFERENCE_ROOM] }
    BookingStatus: { type: string, enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXPIRED] }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        full_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        department_id: { type: string, format: uuid, nullable: true }
        status: { type: string }
        mfa_enabled: { type: boolean }
...
    Room:
      type: object
      properties:
        id: { type: string, format: uuid }
        room_code: { type: string }
        name: { type: string }
        building_id: { type: string, format: uuid }
        capacity: { type: integer }
        room_type: { $ref: "#/components/schemas/RoomType" }
        status: { type: string }
        building_code: { type: string }
        building_name: { type: string }
        image_url: { type: string, nullable: true }
        image_public_id: { type: string, nullable: true }
        gallery_urls: { type: array, items: { type: string } }
        gallery_public_ids: { type: array, items: { type: string } }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        room_id: { type: string, format: uuid }
        requested_by: { type: string, format: uuid }
        purpose: { type: string }
        attendee_count: { type: integer }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/BookingStatus" }
        review_note: { type: string, nullable: true }
        # Optional nested relations, returned on the enriched approvals projection.
        room: { allOf: [ { $ref: "#/components/schemas/Room" } ], nullable: true }
        requester: { allOf: [ { $ref: "#/components/schemas/User" } ], nullable: true }
    ApprovalBlocker:
      type: object
      description: A reason a pending booking can or cannot be approved (§11/FR8).
      required: [kind, message]
      properties:
        kind:
          type: string
          enum: [LECTURE, MAINTENANCE, APPROVED_BOOKING, COMPETING_PENDING, CAPACITY, IN_PAST]
        message: { type: string }
        starts_at: { type: string, format: date-time, nullable: true }
        ends_at: { type: string, format: date-time, nullable: true }
    BookingApprovability:
      type: object
      description: >
        A pending booking enriched with approvability info, used on the approvals
        queue. can_approve is true only when there are no hard blockers (IN_PAST,
        CAPACITY, LECTURE, MAINTENANCE, APPROVED_BOOKING); competing pendings are
        informational and never block.
      required: [booking, can_approve, blockers, competing_pending_count]
      properties:
        booking: { $ref: "#/components/schemas/Booking" }
        can_approve: { type: boolean }
        blockers:
          type: array
          items: { $ref: "#/components/schemas/ApprovalBlocker" }
        competing_pending_count: { type: integer }
...